12/18 라우터 분류
