+++ src/components/Header.js
@@ -16,13 +16,13 @@ function renderLoginStatus() {
 
   if (user.username) {
     return `
-      <li><a href="/" class="${path === "/" ? "text-blue-600" : "text-gray-600"}">홈</a></li>
-      <li><a href="/profile" class="${path === "/profile" ? "text-blue-600" : "text-gray-600"}">프로필</a></li>
+      <li><a href="/" class="${path === "/" ? "text-blue-600 font-bold" : "text-gray-600"}">홈</a></li>
+      <li><a href="/profile" class="${path === "/profile" ? "text-blue-600 font-bold" : "text-gray-600"}">프로필</a></li>
       <li><a href="/login" id="logout" class="text-gray-600">로그아웃</a></li>
     `;
   } else {
     return `
-      <li><a href="/" class="text-blue-600">홈</a></li>
+      <li><a href="/" class="text-blue-600 font-bold">홈</a></li>
       <li><a href="/login" id="login-but" class="text-gray-600">로그인</a></li>
     `;
   }
